# Conflicts:
#	carservice-db/sql/carservice_plus.sql
#	carservice-db/src/main/java/org/click/carservice/db/domain/CarServiceInsuranceInfo.java
#	carservice-db/src/main/java/org/click/carservice/db/domain/CarServiceInsuranceService.java
#	carservice-wx-api/src/main/java/org/click/carservice/wx/service/WxCarService.java
#	carservice-wx-api/src/main/java/org/click/carservice/wx/web/WxCarController.java

+++ carservice-core/src/main/java/org/click/carservice/core/service/CouponVerifyService.java
@@ -66,7 +66,7 @@ public class CouponVerifyService {
 
         // 检查是否超期
         Short timeType = coupon.getTimeType();
-        Integer days = coupon.getDays();
+        Short days = coupon.getDays();
         LocalDateTime now = LocalDateTime.now();
         if (timeType.equals(CouponTimeType.TIME_TYPE_TIME.getStatus())) {
             if (now.isBefore(coupon.getStartTime()) || now.isAfter(coupon.getEndTime())) {
